// src/graphql/types.ts
export type Maybe<T> = T;
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * A date string, such as 2007-12-03, compliant with the ISO 8601 standard for
   * representation of dates and times using the Gregorian calendar.
   **/
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Directory = Node & {
  __typename?: "Directory";
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  sourceInstanceName: Maybe<Scalars["String"]>;
  absolutePath: Maybe<Scalars["String"]>;
  relativePath: Maybe<Scalars["String"]>;
  extension: Maybe<Scalars["String"]>;
  size: Maybe<Scalars["Int"]>;
  prettySize: Maybe<Scalars["String"]>;
  modifiedTime: Maybe<Scalars["Date"]>;
  accessTime: Maybe<Scalars["Date"]>;
  changeTime: Maybe<Scalars["Date"]>;
  birthTime: Maybe<Scalars["Date"]>;
  root: Maybe<Scalars["String"]>;
  dir: Maybe<Scalars["String"]>;
  base: Maybe<Scalars["String"]>;
  ext: Maybe<Scalars["String"]>;
  name: Maybe<Scalars["String"]>;
  relativeDirectory: Maybe<Scalars["String"]>;
  dev: Maybe<Scalars["Int"]>;
  mode: Maybe<Scalars["Int"]>;
  nlink: Maybe<Scalars["Int"]>;
  uid: Maybe<Scalars["Int"]>;
  gid: Maybe<Scalars["Int"]>;
  rdev: Maybe<Scalars["Int"]>;
  blksize: Maybe<Scalars["Int"]>;
  ino: Maybe<Scalars["Int"]>;
  blocks: Maybe<Scalars["Int"]>;
  atimeMs: Maybe<Scalars["Float"]>;
  mtimeMs: Maybe<Scalars["Float"]>;
  ctimeMs: Maybe<Scalars["Float"]>;
  birthtimeMs: Maybe<Scalars["Float"]>;
  atime: Maybe<Scalars["Date"]>;
  mtime: Maybe<Scalars["Date"]>;
  ctime: Maybe<Scalars["Date"]>;
  birthtime: Maybe<Scalars["Date"]>;
};
// ...
export type DirectoryConnection = {
  __typename?: "DirectoryConnection";
  totalCount: Scalars["Int"];
  edges: Array<DirectoryEdge>;
  nodes: Array<Directory>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<DirectoryGroupConnection>;
};
// ...
export type File = Node & {
  __typename?: "File";
  birthtime: Maybe<Scalars["Date"]>;
  birthtimeMs: Maybe<Scalars["Float"]>;
  sourceInstanceName: Maybe<Scalars["String"]>;
  absolutePath: Maybe<Scalars["String"]>;
  relativePath: Maybe<Scalars["String"]>;
  extension: Maybe<Scalars["String"]>;
  size: Maybe<Scalars["Int"]>;
  prettySize: Maybe<Scalars["String"]>;
  modifiedTime: Maybe<Scalars["Date"]>;
  accessTime: Maybe<Scalars["Date"]>;
  changeTime: Maybe<Scalars["Date"]>;
  birthTime: Maybe<Scalars["Date"]>;
  root: Maybe<Scalars["String"]>;
  dir: Maybe<Scalars["String"]>;
  base: Maybe<Scalars["String"]>;
  ext: Maybe<Scalars["String"]>;
  name: Maybe<Scalars["String"]>;
  relativeDirectory: Maybe<Scalars["String"]>;
  dev: Maybe<Scalars["Int"]>;
  mode: Maybe<Scalars["Int"]>;
  nlink: Maybe<Scalars["Int"]>;
  uid: Maybe<Scalars["Int"]>;
  gid: Maybe<Scalars["Int"]>;
  rdev: Maybe<Scalars["Int"]>;
  blksize: Maybe<Scalars["Int"]>;
  ino: Maybe<Scalars["Int"]>;
  blocks: Maybe<Scalars["Int"]>;
  atimeMs: Maybe<Scalars["Float"]>;
  mtimeMs: Maybe<Scalars["Float"]>;
  ctimeMs: Maybe<Scalars["Float"]>;
  atime: Maybe<Scalars["Date"]>;
  mtime: Maybe<Scalars["Date"]>;
  ctime: Maybe<Scalars["Date"]>;
  /** Copy file to static directory and return public url to it */
  publicURL: Maybe<Scalars["String"]>;
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  childImageSharp: Maybe<ImageSharp>;
  childMarkdownRemark: Maybe<MarkdownRemark>;
};
// ...
export type FileConnection = {
  __typename?: "FileConnection";
  totalCount: Scalars["Int"];
  edges: Array<FileEdge>;
  nodes: Array<File>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<FileGroupConnection>;
};

export type FileConnectionDistinctArgs = {
  field: FileFieldsEnum;
};

export type FileConnectionGroupArgs = {
  skip: Maybe<Scalars["Int"]>;
  limit: Maybe<Scalars["Int"]>;
  field: FileFieldsEnum;
};

export type FileEdge = {
  __typename?: "FileEdge";
  next: Maybe<File>;
  node: File;
  previous: Maybe<File>;
};
// ...
export type FileGroupConnection = {
  __typename?: "FileGroupConnection";
  totalCount: Scalars["Int"];
  edges: Array<FileEdge>;
  nodes: Array<File>;
  pageInfo: PageInfo;
  field: Scalars["String"];
  fieldValue: Maybe<Scalars["String"]>;
};
// ...
export type ImageSharp = Node & {
  __typename?: "ImageSharp";
  id: Scalars["ID"];
  fixed: Maybe<ImageSharpFixed>;
  resolutions: Maybe<ImageSharpResolutions>;
  fluid: Maybe<ImageSharpFluid>;
  sizes: Maybe<ImageSharpSizes>;
  original: Maybe<ImageSharpOriginal>;
  resize: Maybe<ImageSharpResize>;
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type ImageSharpConnection = {
  __typename?: "ImageSharpConnection";
  totalCount: Scalars["Int"];
  edges: Array<ImageSharpEdge>;
  nodes: Array<ImageSharp>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<ImageSharpGroupConnection>;
};

export type ImageSharpConnectionDistinctArgs = {
  field: ImageSharpFieldsEnum;
};

export type ImageSharpConnectionGroupArgs = {
  skip: Maybe<Scalars["Int"]>;
  limit: Maybe<Scalars["Int"]>;
  field: ImageSharpFieldsEnum;
};

export type ImageSharpEdge = {
  __typename?: "ImageSharpEdge";
  next: Maybe<ImageSharp>;
  node: ImageSharp;
  previous: Maybe<ImageSharp>;
};
// ...
export type ImageSharpFixed = {
  __typename?: "ImageSharpFixed";
  base64: Maybe<Scalars["String"]>;
  tracedSVG: Maybe<Scalars["String"]>;
  aspectRatio: Maybe<Scalars["Float"]>;
  width: Maybe<Scalars["Float"]>;
  height: Maybe<Scalars["Float"]>;
  src: Maybe<Scalars["String"]>;
  srcSet: Maybe<Scalars["String"]>;
  srcWebp: Maybe<Scalars["String"]>;
  srcSetWebp: Maybe<Scalars["String"]>;
  originalName: Maybe<Scalars["String"]>;
};
// ...
export type ImageSharpFluid = {
  __typename?: "ImageSharpFluid";
  base64: Maybe<Scalars["String"]>;
  tracedSVG: Maybe<Scalars["String"]>;
  aspectRatio: Maybe<Scalars["Float"]>;
  src: Maybe<Scalars["String"]>;
  srcSet: Maybe<Scalars["String"]>;
  srcWebp: Maybe<Scalars["String"]>;
  srcSetWebp: Maybe<Scalars["String"]>;
  sizes: Maybe<Scalars["String"]>;
  originalImg: Maybe<Scalars["String"]>;
  originalName: Maybe<Scalars["String"]>;
  presentationWidth: Maybe<Scalars["Int"]>;
  presentationHeight: Maybe<Scalars["Int"]>;
};
// ...
export type ImageSharpOriginal = {
  __typename?: "ImageSharpOriginal";
  width: Maybe<Scalars["Float"]>;
  height: Maybe<Scalars["Float"]>;
  src: Maybe<Scalars["String"]>;
};
// ...
export type ImageSharpResolutions = {
  __typename?: "ImageSharpResolutions";
  base64: Maybe<Scalars["String"]>;
  tracedSVG: Maybe<Scalars["String"]>;
  aspectRatio: Maybe<Scalars["Float"]>;
  width: Maybe<Scalars["Float"]>;
  height: Maybe<Scalars["Float"]>;
  src: Maybe<Scalars["String"]>;
  srcSet: Maybe<Scalars["String"]>;
  srcWebp: Maybe<Scalars["String"]>;
  srcSetWebp: Maybe<Scalars["String"]>;
  originalName: Maybe<Scalars["String"]>;
};
// ...
export type Internal = {
  __typename?: "Internal";
  content: Maybe<Scalars["String"]>;
  contentDigest: Scalars["String"];
  description: Maybe<Scalars["String"]>;
  fieldOwners: Maybe<Array<Maybe<Scalars["String"]>>>;
  ignoreType: Maybe<Scalars["Boolean"]>;
  mediaType: Maybe<Scalars["String"]>;
  owner: Scalars["String"];
  type: Scalars["String"];
};
// ...
export type MarkdownHeading = {
  __typename?: "MarkdownHeading";
  value: Maybe<Scalars["String"]>;
  depth: Maybe<Scalars["Int"]>;
};

export type MarkdownHeadingFilterInput = {
  value: Maybe<StringQueryOperatorInput>;
  depth: Maybe<IntQueryOperatorInput>;
};

export type MarkdownHeadingFilterListInput = {
  elemMatch: Maybe<MarkdownHeadingFilterInput>;
};

export enum MarkdownHeadingLevels {
  H1 = "h1",
  H2 = "h2",
  H3 = "h3",
  H4 = "h4",
  H5 = "h5",
  H6 = "h6"
}

export type MarkdownRemark = Node & {
  __typename?: "MarkdownRemark";
  id: Scalars["ID"];
  frontmatter: Maybe<MarkdownRemarkFrontmatter>;
  excerpt: Maybe<Scalars["String"]>;
  rawMarkdownBody: Maybe<Scalars["String"]>;
  fileAbsolutePath: Maybe<Scalars["String"]>;
  fields: Maybe<MarkdownRemarkFields>;
  html: Maybe<Scalars["String"]>;
  htmlAst: Maybe<Scalars["JSON"]>;
  excerptAst: Maybe<Scalars["JSON"]>;
  headings: Maybe<Array<Maybe<MarkdownHeading>>>;
  timeToRead: Maybe<Scalars["Int"]>;
  tableOfContents: Maybe<Scalars["String"]>;
  wordCount: Maybe<MarkdownWordCount>;
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type MarkdownRemarkConnection = {
  __typename?: "MarkdownRemarkConnection";
  totalCount: Scalars["Int"];
  edges: Array<MarkdownRemarkEdge>;
  nodes: Array<MarkdownRemark>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<MarkdownRemarkGroupConnection>;
};

export type MarkdownRemarkConnectionDistinctArgs = {
  field: MarkdownRemarkFieldsEnum;
};

export type MarkdownRemarkConnectionGroupArgs = {
  skip: Maybe<Scalars["Int"]>;
  limit: Maybe<Scalars["Int"]>;
  field: MarkdownRemarkFieldsEnum;
};

export type MarkdownRemarkEdge = {
  __typename?: "MarkdownRemarkEdge";
  next: Maybe<MarkdownRemark>;
  node: MarkdownRemark;
  previous: Maybe<MarkdownRemark>;
};

export type MarkdownRemarkFields = {
  __typename?: "MarkdownRemarkFields";
  slug: Maybe<Scalars["String"]>;
};
// ...
export type MarkdownRemarkFrontmatter = {
  __typename?: "MarkdownRemarkFrontmatter";
  title: Maybe<Scalars["String"]>;
  templateKey: Maybe<Scalars["String"]>;
  date: Maybe<Scalars["Date"]>;
  featuredpost: Maybe<Scalars["Boolean"]>;
  featuredimage: Maybe<File>;
  description: Maybe<Scalars["String"]>;
  hero: Maybe<File>;
  heading: Maybe<Scalars["String"]>;
  subheading: Maybe<Scalars["String"]>;
};
// ...
export type MarkdownWordCount = {
  __typename?: "MarkdownWordCount";
  paragraphs: Maybe<Scalars["Int"]>;
  sentences: Maybe<Scalars["Int"]>;
  words: Maybe<Scalars["Int"]>;
};

export type MarkdownWordCountFilterInput = {
  paragraphs: Maybe<IntQueryOperatorInput>;
  sentences: Maybe<IntQueryOperatorInput>;
  words: Maybe<IntQueryOperatorInput>;
};

/** Node Interface */
export type Node = {
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type NodeFilterInput = {
  id: Maybe<StringQueryOperatorInput>;
  parent: Maybe<NodeFilterInput>;
  children: Maybe<NodeFilterListInput>;
  internal: Maybe<InternalFilterInput>;
};

export type NodeFilterListInput = {
  elemMatch: Maybe<NodeFilterInput>;
};

export type PageInfo = {
  __typename?: "PageInfo";
  currentPage: Scalars["Int"];
  hasPreviousPage: Scalars["Boolean"];
  hasNextPage: Scalars["Boolean"];
  itemCount: Scalars["Int"];
  pageCount: Scalars["Int"];
  perPage: Maybe<Scalars["Int"]>;
};
// ...
export type Query = {
  __typename?: "Query";
  file: Maybe<File>;
  allFile: FileConnection;
  imageSharp: Maybe<ImageSharp>;
  allImageSharp: ImageSharpConnection;
  markdownRemark: Maybe<MarkdownRemark>;
  allMarkdownRemark: MarkdownRemarkConnection;
  sitePage: Maybe<SitePage>;
  allSitePage: SitePageConnection;
  sitePlugin: Maybe<SitePlugin>;
  allSitePlugin: SitePluginConnection;
  site: Maybe<Site>;
  allSite: SiteConnection;
  directory: Maybe<Directory>;
  allDirectory: DirectoryConnection;
};
// ...
export type Site = Node & {
  __typename?: "Site";
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  siteMetadata: Maybe<SiteSiteMetadata>;
  port: Maybe<Scalars["Int"]>;
  host: Maybe<Scalars["String"]>;
  polyfill: Maybe<Scalars["Boolean"]>;
  pathPrefix: Maybe<Scalars["String"]>;
  buildTime: Maybe<Scalars["Date"]>;
};

export type SiteBuildTimeArgs = {
  formatString: Maybe<Scalars["String"]>;
  fromNow: Maybe<Scalars["Boolean"]>;
  difference: Maybe<Scalars["String"]>;
  locale: Maybe<Scalars["String"]>;
};

export type SiteConnection = {
  __typename?: "SiteConnection";
  totalCount: Scalars["Int"];
  edges: Array<SiteEdge>;
  nodes: Array<Site>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<SiteGroupConnection>;
};
// ...
export type SitePage = Node & {
  __typename?: "SitePage";
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  path: Maybe<Scalars["String"]>;
  internalComponentName: Maybe<Scalars["String"]>;
  component: Maybe<Scalars["String"]>;
  componentChunkName: Maybe<Scalars["String"]>;
  isCreatedByStatefulCreatePages: Maybe<Scalars["Boolean"]>;
  context: Maybe<SitePageContext>;
  pluginCreator: Maybe<SitePlugin>;
  pluginCreatorId: Maybe<Scalars["String"]>;
  componentPath: Maybe<Scalars["String"]>;
};

export type SitePageConnection = {
  __typename?: "SitePageConnection";
  totalCount: Scalars["Int"];
  edges: Array<SitePageEdge>;
  nodes: Array<SitePage>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<SitePageGroupConnection>;
};
// ...
export type SitePlugin = Node & {
  __typename?: "SitePlugin";
  id: Scalars["ID"];
  parent: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
  resolve: Maybe<Scalars["String"]>;
  name: Maybe<Scalars["String"]>;
  version: Maybe<Scalars["String"]>;
  pluginOptions: Maybe<SitePluginPluginOptions>;
  nodeAPIs: Maybe<Array<Maybe<Scalars["String"]>>>;
  browserAPIs: Maybe<Array<Maybe<Scalars["String"]>>>;
  ssrAPIs: Maybe<Array<Maybe<Scalars["String"]>>>;
  pluginFilepath: Maybe<Scalars["String"]>;
  packageJson: Maybe<SitePluginPackageJson>;
};

export type SitePluginConnection = {
  __typename?: "SitePluginConnection";
  totalCount: Scalars["Int"];
  edges: Array<SitePluginEdge>;
  nodes: Array<SitePlugin>;
  pageInfo: PageInfo;
  distinct: Array<Scalars["String"]>;
  group: Array<SitePluginGroupConnection>;
};